requirements.txt
beautifulsoup4>=4.12.0
html2text>=2020.1.16
lxml>=4.9.0
Pillow>=9.0.0
pytesseract>=0.3.10